Added new parameter to pub for multiple publishes in one call
New parameter to /pub named "multipub".  It is an integer and represents
the ascii code for which delineation to use.  So, if you are sending
new-line delinated data, you would publish to:

/pub?multipub=13

And tab delineated would be

/pub?multipub=9

This argument defaults to 0 representing no delination. (ie: delination
happens at the null character which is, by definition, at the end of
the string)